+++ backend/routes-loader/index.ts
@@ -1,9 +1,9 @@
 import { Application } from "express"
 
 import api from "./api.js"
-import emailFollowupRedirectRoute from "../routes/email-followup-redirect.js"
+import followupRedirectRoute from "../routes/followup-redirect.js"
 
 export function loadRoutes(app: Application): void {
   app.use("/api", api)
-  app.use("/followups", emailFollowupRedirectRoute)
+  app.use("/followups", followupRedirectRoute)
 }
